v0.5.2 - modification to user_recycling_activities table and entity, UserStatsDTO;

+++ src/main/java/io/rewardsapp/dto/UserStatsDTO.java
@@ -2,14 +2,13 @@ package io.rewardsapp.dto;
 
 import lombok.Builder;
 
-import java.math.BigInteger;
-
 @Builder
 public record UserStatsDTO(
-      BigInteger paperRecycled,
-      BigInteger plasticRecycled,
-      BigInteger glassRecycled,
-      BigInteger aluminumRecycled,
-      BigInteger metalsRecycled,
-      BigInteger electronicsRecycled
+      Long paperRecycled,
+      Long plasticRecycled,
+      Long glassRecycled,
+      Long aluminumRecycled,
+      Long metalsRecycled,
+      Long electronicsRecycled
 ) {}
+
